+++ packages/thumbprint-react/components/UIAction/index.jsx
@@ -1,2 +1,0 @@
-export { default as Themed } from './themed.jsx';
-export { default as Plain } from './plain.jsx';